# update idx handling