--- a/idx/memory/memory.go
+++ b/idx/memory/memory.go
@@ -143,7 +143,10 @@ func (m *MemoryIdx) Load(defs []schema.MetricDefinition) int {
 		}
 		m.add(def)
 		// as we are loading the metricDefs from a persistant store, set the lastSave
-		// to the lastUpdate timestamp.
+		// to the lastUpdate timestamp.  This wont exactly match the true lastSave Timstamp,
+		// but it will be close enough and it will always be true that the lastSave was at
+		// or after this time.  For metrics that are sent at or close to real time (the typical
+		// use case), then the value will be within a couple of seconds of the true lastSave.
 		m.DefById[def.Id].LastSave = uint32(def.LastUpdate)
 		num++
 		statMetricsActive.Inc()