- support LastSave property, so we only periodically save to cassandra.
- make adding to the writeQueue a non-blocking operation, unless the
  def has not been updated for 1.5x the updateInterval.

replaces PR #569  and PR #571