@woodsaj 2017-03-21
# Insert benchmarks
## new code
```
$ go test -v -run NONE -bench . -benchtime 1s
BenchmarkIndexing-4   	   30000	     50676 ns/op	    3518 B/op	      61 allocs/op
BenchmarkLoad-4       	  200000	      9338 ns/op	    1745 B/op	      14 allocs/op
```

## current master 
(current benchmarks in master dont work, as they dont write to cassandra. So these results are from local modifications to set updateCassIdx=true)
```
$ go test -v -run NONE -bench . -benchtime 1s
BenchmarkIndexing-4   	   30000	     49439 ns/op	    3518 B/op	      61 allocs/op
BenchmarkLoad-4       	  200000	      9384 ns/op	    1744 B/op	      14 allocs/op
```

Results are pretty much identical because all writes are blocking writes to cassandra.

# updates
## new code
```
$ go test -v -run NONE -bench BenchmarkIndexingWithUpdates
BenchmarkIndexingWithUpdates-4   	  500000	      2376 ns/op	     760 B/op	      14 allocs/op
```

## current master
```
$ go test -v -run NONE -bench BenchmarkIndexingWithUpdates
BenchmarkIndexingWithUpdates-4   	   30000	     48469 ns/op	    2693 B/op	      51 allocs/op
```

When performing updates that dont require a save to cassandra, the new code is way faster.