@Dieterbe 2017-03-21
> replaces PR #569 and PR #571

571 also had a change to make the index Pruning use LastSave instead of LastUpdate. I think we should still apply that, as it seems to make more sense