@woodsaj 2017-03-21
LastSave cant be used for pruning, as not all nodes save metricDefs.  So if updateCassIdx=false, then LastSave always be what was set when the def was loaded at startup.