@Dieterbe 2017-03-21
for nodes that don't write to cassandra, we could just make it so that LastSave means "last save to the memory index" instead of "last save to the cassandra index". in other words, just have it be the timestamp of when it was last seen. that way we could do this.  But anyway I don't feel strongly about, I don't have a good example to make a strong case for this (other than people for some reason loading in old data, which doesn't seem very common), so it's fine.