@woodsaj 2017-03-21
for hosted-metrics we the default MAX_STALE is 48hours. So users would have be streaming data with a delay of 48hours (as LastUpdate is now set for every point received instead of the previous periodic saving of lastUpdate).   It is extremely unlikely anyone will send data with that much lag, and if they do we can just change MAX_STALE or turn pruning off completely.