@Dieterbe 2017-03-21
btw to compare benchmarks, these tools are nice:
https://godoc.org/golang.org/x/tools/cmd/benchcmp
https://godoc.org/golang.org/x/perf/cmd/benchstat